Initial Insertion
This program contains all of the different types of paper you can buy. The measurements have millimeters and inches. The program also converts the millimeter entries to inches (ie: X / 25.4) and saves the resulting values to show the difference between how it is shown as inches versus the computed value.